Streamlined handling inner result enum for operator functions
A function was made that is now called at the end of most operator
functions since it's the same approach of checking the result enum,
pushing a value to state and returning Ok if it's good and Err if
it's an error. This avoids a ton of repeated code and is more
understandable.