Added stringCompare operator which compares the contents of strings
Unlike comparing string boxes using the standard comparison operators
of ==, <, >, etc, this operator compares the actual contents of
the string boxes with each other rather than just comparing the
string box numbers. This is useful for seeing if the contents of
a string box are equal to another string box contents rather than
just seeing if the string box points to the same contents as the
other one.