Greatly sped up program by changing up operator calling
Instead of a hashmap, an array indexing system is now used where
an operator word token stores the array index associated with
the operator if it's valid. This removes the former hashing
overhead and more than doubles the program's speed even in the
interpreter at debug-level!

The code involved still definitely needs some improvements
with some cursed code starting on lines 3974 and 4266.
Basically some streamlining needs to be done to make them less
cringe to look at.