Removed arbitrary constants from alias adding in lexxing function
Instead of integers plugged in that match the indicies of the
operators that the aliases are aliasing for, the number for the
original operator is queried from the hashmap directly, which
ties them together.

Now, as long as the order of functions is preserved for both
the function pointer vec and the hashmap in the lex function,
new operators can easily be added.