Reworked format of debugPrintHeap to be more readable
The number of free'd boxes and the total number of things in the
heap have been moved to be the bottom two items of the print
so that way if the heap is huge, the user can still see the total
counts without a bunch of stuff getting in the way.

Neatly, this also acts as a built-in Valgrind, with the user being
able to check the ratio of things on the heap to things free'd on
the heap by comparing the two counts. Equality means all is free'd,
inequality means some items still need to be free'd.

The terminology by each memory cell was also changed to say [FREE]
instead of [INVALID] since [INVALID] implies the cell can't be used
which is the exact opposite of the intended indication.
[FREE] directly says to the user that the cell is free to use.

Keeping it saying invalid for debugPrintStack makes sense because
in that case those boxes that say that are indeed invalid and must
not be used or an error happens.