Added basic format for README.md with table of contents
The README now has sections created with filler Lorem Ipsum text
filled in them to pad it out and to test the links.
The table of contents allows the user to jump to any section
of the document and each section has a link on the bottom that
allows the user to return to the table of contents.