Added ability to run a program from stdin in Lmao
Instead of an error when no file arguments are given
on the command line, a prompt is given for user input to write
a program.

Once Ctrl-D is pressed, typically twice, stdin closes and the
program is read in, interpreted, and run like it would be if it
came from a file stored on disk.

This is useful for rapid demonstration of the programming language's
functionality but also can be nice for whipping up a quick test
of a new feature in the language.

Generally, it's pretty nice. It's still far off from a REPL though.