Added more lenient parsing
If the user now types a data type on its own such as `f32` `isize`
`u128`, etc, there won't be an error saying it's an incorrect
value of the indicated type, instead just an unknown operator.
This is so castTo can have the datatype typed next to it and not
as a string. In other words, you'll be able to type
`castTo u32 ;` instead of needing to type `castTo "u32" ;`.