Added means of multiple import protection
This completes at least the basic form of the import() decerator.
Of course, more must be done to make the detection robust.
For instance, it's possible to get to the same file through different
paths which is not accounted for as of yet. However, it works
well enough for general use cases, with the edge cases being
something to address in due time.